formalised html layout, direct message starts, both untested yet

+++ routes.py
@@ -2,10 +2,11 @@ from flask import redirect, render_template, request, session, abort
 import readdb
 import writedb
 from app import app
+import secrets
 
 
-#Käyttäjä keskeiset sivut
-#User centric pages
+#Käyttäjän luonti ja kirjautuminen
+#User creation and login
 
 
 @app.route("/account")
@@ -20,13 +21,14 @@ def newaccount(error=False, error_msg=""):
 def create():
     name = request.form["accname"]
     password = request.form["passwrd"]
-    if password == None or name == None or name == "" or password == "":
+    if not password or not name:
         return newaccount(True, "Nimi tai salasanakenttä tyhjä")
     #Tähän sitten salasanan laatutarkastus
     #elif (len(password) < 8) or (name in password):
         #return "<script>alert('Kelvoton salasana')</script><script>document.location='/newaccount'</script>"
     if writedb.add_user(name, password):
         session["username"] = name
+        session["csrf_token"] = secrets.token_hex(16)
         return redirect("/")
     return newaccount(True, "Käyttäjänimi varattu")
 
@@ -36,13 +38,52 @@ def login():
     password = request.form["passwrd"]    
     if readdb.confirm_user_pass(name, password):
         session["username"] = name
+        session["csrf_token"] = secrets.token_hex(16)
         return redirect("/")
     return account(True, "Väärä nimi tai salasana")
+
+
+#Uloskirjautuminen ja käyttäjän poistaminen
+#Logout and user removal
+
 
 @app.route("/logout")
 def logout():
     del session["username"]
+    del session["csrf_token"]
     return redirect("/")
+
+@app.route("/changepass")
+def changepass(error=False, error_msg=""):
+    return render_template("changepass.html", error=error, error_msg=error_msg)
+
+@app.route("/changepass/confirm", methods=["POST"])
+def changeconfirm():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    if readdb.confirm_user_pass(session["username"], request.form["oldpasswrd"]) and not (request.form["newpasswrd"] == None or request.form["newpasswrd"] == ""):
+        writedb.change_user_pass(session["username"], request.form["newpasswrd"])
+        return redirect("/personal")
+    else:
+        return changepass(True, "väärä salasana")
+
+@app.route("/rmvuser")
+def rmvuser():
+    return render_template("rmvuser.html")
+
+@app.route("/rmvuser/confirm", methods=["POST"])
+def rmvconfirm():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    if readdb.confirm_user_pass(request.form["accname"], request.form["passwrd"]):
+        writedb.delete_user(request.form["accname"])
+        del session["username"]
+        return redirect("/")
+
+
+#Kaverilista ja muut käyttäjäkohtaiset
+#Friendlist and other user oriented stuff
+
 
 @app.route("/personal")
 def personal(msg=False, msg_content=""):
@@ -60,10 +101,14 @@ def friends(request=False, req_msg=""):
 
 @app.route("/friendlist/send")
 def friendsend(request=False, req_msg=""):
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     return render_template("friendsend.html", request=request, req_msg=req_msg)
 
 @app.route("/friendlist/send/confirm", methods=["POST"])
 def sendfriend():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     if readdb.check_user_exist(request.form["friend"]):
         writedb.add_friend(session["username"], request.form["friend"])
         return friends(True, "Kaverikutsu lähetetty")
@@ -74,29 +119,34 @@ def sentrequests():
     requested = readdb.get_friend_requests(session["username"])
     return render_template("requests.html", requested=requested)
 
-@app.route("/changepass")
-def changepass(error=False, error_msg=""):
-    return render_template("changepass.html", error=error, error_msg=error_msg)
+@app.route("/friendlist/accept")
+def sentrequests():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    writedb.add_friend(session["username"],request.form["username"])
+    return redirect("/friendlist/received")
 
-@app.route("/changepass/confirm", methods=["POST"])
-def changeconfirm():
-    if readdb.confirm_user_pass(session["username"], request.form["oldpasswrd"]) and not (request.form["newpasswrd"] == None or request.form["newpasswrd"] == ""):
-        writedb.change_user_pass(session["username"], request.form["newpasswrd"])
-        return redirect("/personal")
-    else:
-        return changepass(True, "väärä salasana")
+@app.route("/friendlist/reject")
+def sentrequests():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    writedb.remove_friend(request.form["username"],session["username"])
+    return redirect("/friendlist/received")
 
-@app.route("/rmvuser")
-def rmvuser():
-    return render_template("rmvuser.html")
+@app.route("/friendlist/dm")
+def directmessage():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    requested = readdb.get_direct_messages(session["username"], request.form["username"])
+    return render_template("directmessage.html", requested, friend=request.form["username"])
 
-@app.route("/rmvuser/confirm", methods=["POST"])
-def rmvconfirm():
-    if readdb.confirm_user_pass(request.form["accname"], request.form["passwrd"]):
-        writedb.delete_user(request.form["accname"])
-        del session["username"]
-        return redirect("/")
+@app.route("/friendlist/dm/search")
 
+
+@app.route("/friendlist/dm/msg-delete<int:id>")
+
+
+@app.route("/friendlist/dm/newmessage")
 
 #hakutoiminnot
 #search functionality
@@ -160,16 +210,12 @@ def discussion(id1, id2, error=False, error_msg=""):
 
 @app.route("/newarea")
 def newarea(error=False, error_msg=""):
-    is_admin = False
-    if session["username"] != None:
-        is_admin = readdb.is_admin(session["username"])
-    if is_admin:
-        return render_template("newarea.html", error=error, error_msg=error_msg)
-    else:
-        abort(403)
+    return render_template("newarea.html", error=error, error_msg=error_msg)
 
 @app.route("/newarea/confirm", methods=["POST"])
 def newareaconfirm():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     is_admin = False
     if session["username"] != None:
         is_admin = readdb.is_admin(session["username"])
@@ -185,6 +231,8 @@ def newdiscussion(id, error=False, error_msg=""):
 
 @app.route("/area/<int:id>/creatediscussion", methods=["POST"])
 def creatediscussion(id):
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     title = request.form["title"]
     message = request.form["body"]
     if title == None or message == None or title == "" or message == "":
@@ -196,6 +244,8 @@ def creatediscussion(id):
 
 @app.route("/area/<int:id1>/discussion/<int:id2>/newmessage", methods=["POST"])
 def newmessage(id1, id2):
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     message = request.form["bodyfield"]
     if message == None:
         return discussion(id1=id1, id2=id2, error=True, error_msg="Viestikenttä on tyhjä")
@@ -209,6 +259,8 @@ def newmessage(id1, id2):
 
 @app.route("/area/<int:id1>/discussion/<int:id2>/msg-delete/<int:id3>", methods=["POST"])
 def msgDelete(id1, id2, id3):
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     if readdb.get_message_by_id(id3).writer == session["username"]:
         writedb.delete_message(id3)
     else:
